:truck: Rename SummaryHeaderPanel to SummaryHeaderPanelLegacy.
Why? To make space for a newer style of travel authorization summary
panel.
