+++ web/src/components/travel-authorizations/SummaryHeaderPanelLegacy.vue
